refactor(insim_extra): fold Presence and Game into World
