+++ examples/clockwork-carnage/src/games/metronome/handlers.rs
@@ -40,11 +40,11 @@ fn start_setup(state: &State<Metronome>, world: &World, sender: &Sender, ui: &Me
     ));
     refresh_ui(state, ui);
 
-    let game = world.game().clone();
+    let world = world.clone();
     let sender = sender.clone();
     drop(tokio::spawn(async move {
         let result = tokio::select! {
-            r = track_rotation(&game, track, RaceLaps::Untimed, 0, layout, setup_cancel.clone(), &sender) => r,
+            r = track_rotation(&world, track, RaceLaps::Untimed, 0, layout, setup_cancel.clone(), &sender) => r,
             _ = tokio::time::sleep(setup_timeout) => None,
         };
         match result {
@@ -231,10 +231,7 @@ pub(super) async fn on_uco(
     if player.ptype.contains(insim::insim::PlayerType::AI) {
         return Ok(());
     }
-    let uname = world
-        .get(player.ucid)
-        .map(|c| c.uname)
-        .unwrap_or_default();
+    let uname = world.get(player.ucid).map(|c| c.uname).unwrap_or_default();
 
     if is_cp1 {
         let start_time = uco.time;
