+++ examples/clockwork-carnage/src/games/shortcut/handlers.rs
@@ -38,11 +38,11 @@ fn start_setup(state: &State<Shortcut>, world: &World, sender: &Sender, ui: &Sho
     ));
     refresh_ui(state, ui);
 
-    let game = world.game().clone();
+    let world = world.clone();
     let sender = sender.clone();
     drop(tokio::spawn(async move {
         let result = tokio::select! {
-            r = track_rotation(&game, track, RaceLaps::Untimed, 0, layout, setup_cancel.clone(), &sender) => r,
+            r = track_rotation(&world, track, RaceLaps::Untimed, 0, layout, setup_cancel.clone(), &sender) => r,
             _ = tokio::time::sleep(setup_timeout) => None,
         };
         match result {
@@ -228,10 +228,7 @@ pub(super) async fn on_uco(
     if player.ptype.contains(PlayerType::AI) {
         return Ok(());
     }
-    let uname = world
-        .get(player.ucid)
-        .map(|c| c.uname)
-        .unwrap_or_default();
+    let uname = world.get(player.ucid).map(|c| c.uname).unwrap_or_default();
 
     if is_cp1 {
         let _ = state
