+++ kitcar/src/presence.rs
@@ -1,13 +1,6 @@
-//! [`Handler`] and [`FromContext`] impls for [`insim_extra::presence::Presence`],
-//! plus the synthetic event structs that `kitcar` dispatch uses.
+//! Synthetic event structs for presence changes, re-exported from `insim_extra`.
 
-use std::future::Future;
-
-use insim::WithRequestId;
-pub use insim_extra::presence::{ConnectionInfo, PlayerInfo, Presence, PresenceEvent};
-use insim_extra::world::World;
-
-use crate::{AppError, Dispatch, ExtractCx, FromContext, Handler, Sender, Startup};
+pub use insim_extra::presence::{ConnectionInfo, PlayerInfo};
 
 /// Synthetic event emitted when a connection joins.
 #[derive(Debug, Clone)]
@@ -66,70 +59,3 @@ pub struct TakingOver {
 /// Synthetic event emitted when a player tele-pits (Shift+P).
 #[derive(Debug, Clone)]
 pub struct PlayerTeleportedToPits(pub PlayerInfo);
-
-/// Extract [`Presence`] from a registered [`World`].
-impl<S> FromContext<S> for Presence {
-    fn from_context(cx: &ExtractCx<'_, S>) -> Option<Self> {
-        cx.lookup::<World>().map(|w| w.presence().clone())
-    }
-}
-
-/// [`Handler`] impl delegates to [`Presence::apply_packet`] and emits each
-/// change as a typed synthetic event. Register at [`crate::Stage::Pre`] so the
-/// connection / player maps are settled before any Update-stage handler reads them.
-///
-/// On [`Startup`] it also sends `Tiny::Ncn` and `Tiny::Npl` to request the
-/// existing connection and player lists from LFS (they are not sent automatically
-/// on connect).
-impl<S: Send + Sync + 'static> Handler<(), S> for Presence {
-    fn call(self, cx: &ExtractCx<'_, S>) -> impl Future<Output = Result<(), AppError>> + Send {
-        let events = if let Dispatch::Packet(p) = cx.dispatch {
-            self.apply_packet(p)
-        } else {
-            vec![]
-        };
-        let startup = if let Dispatch::Synthetic(s) = cx.dispatch {
-            s.downcast_ref::<Startup>().is_some()
-        } else {
-            false
-        };
-        let sender = cx.sender.clone();
-        async move {
-            if startup {
-                for t in Presence::STARTUP_REQUESTS {
-                    let _ = sender.packet(t.clone().with_request_id(1));
-                }
-            }
-            emit_presence_events(events, &sender);
-            Ok(())
-        }
-    }
-}
-
-fn emit_presence_events(events: Vec<PresenceEvent>, sender: &Sender) {
-    for event in events {
-        let _ = match event {
-            PresenceEvent::Connected(info) => sender.event(Connected(info)),
-            PresenceEvent::Disconnected { ucid, info } => sender.event(Disconnected { ucid, info }),
-            PresenceEvent::ConnectionDetails(info) => sender.event(ConnectionDetails(info)),
-            PresenceEvent::VehicleSelected { ucid, vehicle } => {
-                sender.event(VehicleSelected { ucid, vehicle })
-            },
-            PresenceEvent::Renamed {
-                ucid,
-                uname,
-                new_pname,
-            } => sender.event(Renamed {
-                ucid,
-                uname,
-                new_pname,
-            }),
-            PresenceEvent::PlayerJoined(p) => sender.event(PlayerJoined(p)),
-            PresenceEvent::PlayerLeft(p) => sender.event(PlayerLeft(p)),
-            PresenceEvent::TakingOver { before, after } => {
-                sender.event(TakingOver { before, after })
-            },
-            PresenceEvent::PlayerTeleportedToPits(p) => sender.event(PlayerTeleportedToPits(p)),
-        };
-    }
-}
